+++ mixer/bl_operators.py
@@ -24,8 +24,10 @@ import os
 import socket
 import subprocess
 import time
+from pathlib import Path
 
 import bpy
+from bpy_extras.io_utils import ImportHelper
 
 from mixer.share_data import share_data
 from mixer.bl_utils import get_mixer_props, get_mixer_prefs
@@ -43,7 +45,57 @@ logger = logging.getLogger(__name__)
 
 
 poll_is_client_connected = (lambda: is_client_connected(), "Client not connected")
-poll_already_in_a_room = (lambda: not share_data.client.current_room, "Already in a room")
+poll_already_in_a_room = (lambda: not is_client_connected() or not share_data.client.current_room, "Already in a room")
+
+
+class SharedFoldersAddFolderOperator(bpy.types.Operator, ImportHelper):
+    bl_idname = "mixer.add_shared_folder"
+    bl_label = "Add Shared Folder"
+    filepath: bpy.props.StringProperty(subtype="DIR_PATH")
+
+    def execute(self, context):
+
+        path = self.filepath
+        if not Path(path).is_dir():
+            path = str(Path(path).parent)
+
+        for item in get_mixer_prefs().shared_folders:
+            if item.shared_folder == path:
+                return {"FINISHED"}
+
+        item = get_mixer_prefs().shared_folders.add()
+        item.shared_folder = path
+        return {"FINISHED"}
+
+    @classmethod
+    def poll_functors(cls, context):
+        return [
+            poll_already_in_a_room,
+        ]
+
+    @classmethod
+    def poll(cls, context):
+        return generic_poll(cls, context)
+
+
+class SharedFoldersRemoveFolderOperator(bpy.types.Operator):
+    bl_idname = "mixer.remove_shared_folder"
+    bl_label = "Remove Shared Folder"
+
+    def execute(self, context):
+        props = get_mixer_props()
+        get_mixer_prefs().shared_folders.remove(props.shared_folder_index)
+        return {"FINISHED"}
+
+    @classmethod
+    def poll_functors(cls, context):
+        return [
+            poll_already_in_a_room,
+        ]
+
+    @classmethod
+    def poll(cls, context):
+        return generic_poll(cls, context)
 
 
 def generic_poll(cls, context):
@@ -91,10 +143,14 @@ class CreateRoomOperator(bpy.types.Operator):
         if not is_client_connected():
             return {"CANCELLED"}
 
-        prefs = get_mixer_prefs()
-        room = prefs.room
+        mixer_prefs = get_mixer_prefs()
+        room = mixer_prefs.room
         logger.warning(f"CreateRoomOperator.execute({room})")
-        join_room(room, prefs.vrtist_protocol)
+
+        shared_folders = []
+        for item in mixer_prefs.shared_folders:
+            shared_folders.append(item.shared_folder)
+        join_room(room, mixer_prefs.vrtist_protocol, shared_folders)
 
         return {"FINISHED"}
 
@@ -151,8 +207,11 @@ class JoinRoomOperator(bpy.types.Operator):
         room = props.rooms[room_index].name
         logger.warning(f"JoinRoomOperator.execute({room})")
 
-        prefs = get_mixer_prefs()
-        join_room(room, prefs.vrtist_protocol)
+        mixer_prefs = get_mixer_prefs()
+        shared_folders = []
+        for item in mixer_prefs.shared_folders:
+            shared_folders.append(item.shared_folder)
+        join_room(room, mixer_prefs.vrtist_protocol, shared_folders)
 
         return {"FINISHED"}
 
@@ -356,7 +415,10 @@ class LaunchVRtistOperator(bpy.types.Operator):
                 return {"CANCELLED"}
 
             logger.warning("LaunchVRtistOperator.execute({mixer_prefs.room})")
-            join_room(mixer_prefs.room, True)
+            shared_folders = []
+            for item in mixer_prefs.shared_folders:
+                shared_folders.append(item.shared_folder)
+            join_room(mixer_prefs.room, True, shared_folders)
 
             # Wait for room creation/join
             timeout = 10
@@ -417,6 +479,8 @@ classes = (
     LeaveRoomOperator,
     DownloadRoomOperator,
     UploadRoomOperator,
+    SharedFoldersAddFolderOperator,
+    SharedFoldersRemoveFolderOperator,
 )
 
 register_factory, unregister_factory = bpy.utils.register_classes_factory(classes)