+++ mixer/bl_preferences.py
@@ -47,6 +47,10 @@ def set_log_level(self, value):
     logger.log(value, "Logging level changed")
 
 
+class SharedFolderItem(bpy.types.PropertyGroup):
+    shared_folder: bpy.props.StringProperty(default="", subtype="DIR_PATH", name="Shared Folder")
+
+
 class MixerPreferences(bpy.types.AddonPreferences):
     """
     Preferences class, store persistent properties and options.
@@ -124,6 +128,8 @@ class MixerPreferences(bpy.types.AddonPreferences):
         name="Data Directory", default=os.environ.get("MIXER_DATA_DIR", get_data_directory()), subtype="FILE_PATH"
     )
 
+    shared_folders: bpy.props.CollectionProperty(name="Shared Folders", type=SharedFolderItem)
+
     # Developer option to avoid sending scene content to server at the first connexion
     # Allow to quickly iterate debugging/test on large scenes with only one client in room
     # Main usage: optimization of client timers to check if updates are required
@@ -150,7 +156,10 @@ class MixerPreferences(bpy.types.AddonPreferences):
         draw_preferences_ui(self, context)
 
 
-classes = (MixerPreferences,)
+classes = (
+    SharedFolderItem,
+    MixerPreferences,
+)
 
 register_factory, unregister_factory = bpy.utils.register_classes_factory(classes)