+++ mixer/connection.py
@@ -48,7 +48,7 @@ def set_client_attributes():
     )
 
 
-def join_room(room_name: str, vrtist_protocol: bool = False):
+def join_room(room_name: str, vrtist_protocol: bool = False, shared_folders=None):
     prefs = get_mixer_prefs()
     logger.warning(f"join: room: {room_name}, user: {prefs.user}")
 
@@ -64,7 +64,9 @@ def join_room(room_name: str, vrtist_protocol: bool = False):
     set_client_attributes()
     share_data.client.join_room(room_name)
 
-    share_data.set_vrtist_protocol(vrtist_protocol)
+    if shared_folders is None:
+        shared_folders = []
+    share_data.init_protocol(vrtist_protocol, shared_folders)
     share_data.pending_test_update = False
 
     # join a room <==> want to track local changes