+++ tests/blender_app.py
@@ -25,17 +25,14 @@ bpy.ops.mixer.disconnect()
 
 _create_room = """
 from mixer.connection import join_room
-join_room("{room_name}", {vrtist_protocol})
+join_room("{room_name}", {vrtist_protocol}, {shared_folders})
 """
 
 _keep_room_open = """
 from mixer.share_data import share_data
 share_data.client.set_room_keep_open("{room_name}", {keep_room_open})
 """
-_set_workspace = """
-from mixer.tructruc import set_workspace
-set_workspace({workspace_folders})
-"""
+
 _join_room = """
 from mixer.connection import join_room
 from mixer.broadcaster.common import RoomAttributes
@@ -63,7 +60,7 @@ def wait_joinable():
     return room_attributes is not None and room_attributes.get(RoomAttributes.JOINABLE, False)
 
 if wait_joinable():
-    join_room("{room_name}", {vrtist_protocol})
+    join_room("{room_name}", {vrtist_protocol}, {shared_folders})
 else:
     print(f"ERROR: Cannot join room after {max_wait} seconds. Abort")
     time.sleep(5)
@@ -99,14 +96,12 @@ class BlenderApp:
         room_name="mixer_unittest",
         keep_room_open=False,
         vrtist_protocol: bool = False,
-        workspace_folders: List[str] = (),
+        shared_folders: List[str] = (),
     ):
         """Emit a mixer create room command"""
-        if workspace_folders:
-            set_workspaces = _set_workspace.format(workspace_folders=list(workspace_folders))
-            self._blender.send_string(set_workspaces)
-
-        create_room = _create_room.format(room_name=room_name, vrtist_protocol=vrtist_protocol)
+        create_room = _create_room.format(
+            room_name=room_name, vrtist_protocol=vrtist_protocol, shared_folders=list(shared_folders)
+        )
         self._blender.send_string(create_room)
 
         keep_room_open = _keep_room_open.format(room_name=room_name, keep_room_open=keep_room_open)
@@ -116,13 +111,15 @@ class BlenderApp:
         self,
         room_name="mixer_unittest",
         vrtist_protocol: bool = False,
-        workspace_folders: Iterable[str] = (),
+        shared_folders: Iterable[str] = (),
     ):
         """Emit a mixer join room command"""
-        if workspace_folders:
-            set_workspaces = _set_workspace.format(workspace_folders=list(workspace_folders))
-            self._blender.send_string(set_workspaces)
-        join_room = _join_room.format(room_name=room_name, vrtist_protocol=vrtist_protocol, max_wait="{max_wait}")
+        join_room = _join_room.format(
+            room_name=room_name,
+            vrtist_protocol=vrtist_protocol,
+            shared_folders=list(shared_folders),
+            max_wait="{max_wait}",
+        )
         self._blender.send_string(join_room)
 
     def disconnect_mixer(self):